getting closer to working

+++ src/usart.cpp
@@ -24,7 +24,7 @@ USART Serial;
 
 // when they said fractional baud rate i didnt think they literally meant it
 // http://ww1.microchip.com/downloads/en/AppNotes/TB3216-Getting-Started-with-USART-90003216A.pdf
-#define USART0_BAUD_RATE(r) ((float)(F_CPU * 64 / (16 * (float)(r))) + 0.5)
+#define USART0_BAUD_RATE(r) ((float)(((float)F_CPU) * 64 / (16 * (float)(r))) + 0.5)
 
 void USART::begin() {
   // set PORTB.PIN2 (TXD) as output
